# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Teslog API
  description: >
    Self-hosted Tesla vehicle data logging platform. Tracks drives, charges,
    idle sessions, battery health, and firmware updates via Tesla Fleet
    Telemetry.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: "{server_url}/api"
    variables:
      server_url:
        default: http://localhost
        description: Base URL of the Teslog instance
...
security:
  - bearerAuth: []
paths:
...
  /v1/auth/register:
    post:
      tags: [Auth]
      summary: Register the first user (first-run only)
      security: []
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, password_confirmation]
              properties:
                name:
                  type: string
                  maxLength: 255
                email:
                  type: string
                  format: email
                  maxLength: 255
                password:
                  type: string
                  format: password
                password_confirmation:
                  type: string
                  format: password
                timezone:
                  type: string
                  example: America/New_York
                distance_unit:
                  type: string
                  enum: [mi, km]
                temperature_unit:
                  type: string
                  enum: [F, C]
                currency:
                  type: string
                  maxLength: 3
                  example: USD
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  token:
                    type: string
                    description: Sanctum plain-text token
        "403":
          description: Registration closed (user already exists)
        "422":
          $ref: "#/components/responses/ValidationError"
  /v1/auth/login:
    post:
      tags: [Auth]
      summary: Login and obtain an API token
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                device_name:
                  type: string
                  description: Label for the token (default "api")
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  token:
                    type: string
        "401":
          description: Invalid credentials
        "429":
          description: Too many attempts (rate limited to 5 per minute)
...
  /v1/auth/tokens/{tokenId}:
    delete:
      tags: [Auth]
      summary: Revoke a specific token
      operationId: deleteToken
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Token revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Vehicles ────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Sanctum API token obtained from /v1/auth/login or /v1/auth/register
    telemetryAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Shared secret for fleet-telemetry ingestion
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
              errors:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
  schemas:
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        timezone:
          type: string
        distance_unit:
          type: string
          enum: [mi, km]
        temperature_unit:
          type: string
          enum: [F, C]
        elevation_unit:
          type: string
          enum: [ft, m]
        currency:
          type: string
        theme:
          type: string
          nullable: true
        debug_mode:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...